Move buffer tests into their own file
