+++ crates/buffer/src/tests.rs
@@ -1,0 +1,2 @@
+mod buffer;
+mod syntax;
